+++ Libplanet.Tests/Blocks/BlockMetadataExtensionsTest.cs
@@ -1,0 +1,25 @@
+using System;
+using Libplanet.Blocks;
+using Xunit;
+
+namespace Libplanet.Tests.Blocks
+{
+    public class BlockMetadataExtensionsTest
+    {
+        [Fact]
+        public void ValidateTimestamp()
+        {
+            DateTimeOffset now = DateTimeOffset.UtcNow;
+            DateTimeOffset future = now + TimeSpan.FromSeconds(17);
+            IBlockMetadata metadata = new BlockMetadata
+            {
+                Timestamp = future,
+            };
+
+            Assert.Throws<InvalidBlockTimestampException>(() => metadata.ValidateTimestamp(now));
+
+            // It's okay because 3 seconds later.
+            metadata.ValidateTimestamp(now + TimeSpan.FromSeconds(3));
+        }
+    }
+}